--- client/src/api/useAuthContext.js
@@ -1,9 +1,0 @@
-import AuthContext from "../auth/AuthContext";
-import { useContext } from "react";
-export const useAuthContext = () => {
-    const user = useContext(AuthContext);
-    if (user === undefined) {
-        throw new Error("useAuthContext can only be used inside AuthProvider");
-    }
-    return user;
-};
