+++ client/src/core/auth/IsAuthenticated.jsx
@@ -1,0 +1,20 @@
+import { useNavigate } from "react-router-dom";
+import { useAuthContext } from "./useAuthContext";
+import { useEffect } from "react";
+import Loading from "../../components/Common/Loading";
+
+export const IsAuthenticated = () => {
+  const navigate = useNavigate();
+  const context = useAuthContext();
+
+  useEffect(() => {
+    if (context?.user?.isAuthenticated) {
+      console.log("User is authenticated, redirecting to overview..");
+      navigate("/overview");
+    } else {
+      navigate("/login");
+    }
+  }, [context?.user?.isAuthenticated, navigate]);
+
+  return <Loading />;
+};
