+++ client/src/index.js
@@ -4,7 +4,7 @@ import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
-import { AuthProvider } from "./auth/AuthProvider";
+import { AuthProvider } from "./core/auth/AuthProvider";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
